delete testNewPipelineItemDisplayedOnDashboard() as it is duplicate of testCreatePipelineAssertOnDashboard() move testCreatePipelineExistingNameError() from PipelineTest to NewItemTest move testCreateNewPipeline() from PipelineTest to NewItemTest and rename into testCreatePipelineAssertInsideJob() move testCreatePipelineWithName() from PipelineTest to NewItemTest and rename into testCreatePipelineAssertOnDashboard() move testCreateNewPipelineWithDescription() from PipelineTest to NewItemTest move testCreateNewPipelineFromExisting() from PipelineTest to NewItemTest move testCreatedPipelineIsDisplayedOnBreadcrumbs() from PipelineTest to NewItemTest and rename into testCreatePipelineAssertOnBreadcrumbs rename testPipelineAEditDescription() into testEditPipelineDescription() delete testPipelineAEditDescription as it is duplicate of testEditPipelineDescription()
ignore testAddingGitRepository()
ignore testEditPipelineDescription()
ignore testEnablePipelineProject()
ignore testPipelineSideMenuLinks()
ignore testPipelineSideMenuLinks()
ignore testWarningMessageIsDisappeared()
